+++ previous-versions/common/RELEASES.md
@@ -681,3 +681,13 @@ Based on:
 - [typescript v0.38.0] previous-versions/common
 ### Releases
 - [NPM v0.38.0] https://www.npmjs.com/package/@codat/common/v/0.38.0 - previous-versions/common
+
+## 2023-11-06 01:29:42
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.112.1 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.39.0] previous-versions/common
+### Releases
+- [NPM v0.39.0] https://www.npmjs.com/package/@codat/common/v/0.39.0 - previous-versions/common